backend/requirements.txt
asyncio==3.4.3
blinker==1.8.2
bson==0.5.10
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-OAuthlib==0.9.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==2.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
pandas==2.0.3
requests==2.32.2
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3